feat: implement loading of dreams

+++ app/src/main/java/app/dreamjournal/data/Constants.kt
@@ -1,3 +1,3 @@
 package app.dreamjournal.data
 
-const val ID_NOT_SET: Int = 0
+const val ID_NOT_SET: Long = 0
