+++ app/src/main/java/app/dreamjournal/data/dream/Dream.kt
@@ -16,7 +16,7 @@ data class Dream(
     val isLucid: Boolean = false,
     val isFavorite: Boolean = false,
     val color: TagColor = TagColor.White,
-    @PrimaryKey(autoGenerate = true) val id: Int = ID_NOT_SET,
+    @PrimaryKey(autoGenerate = true) val id: Long = ID_NOT_SET,
 ) {
     @Ignore
     val isSaved = id != ID_NOT_SET
